add  detection button and functionality to the dashboard
