+++ interface/src/App.js
@@ -1,22 +1,24 @@
-import React from 'react';
+import React, { Component } from 'react'
 import './App.css';
 import Header from './components/Header'
 import Dashboard from './components/Dashboard';
 import CamPage from './components/CamPage';
 import { Router } from '@reach/router'
 import LoginPage from './components/LoginPage';
+import {startVideo} from './utils/utils'
 
-function App() {
-  return (
-    <div className="App">
-      <Header />
-      <Router>
-        <LoginPage path='/' />
-        <Dashboard path='/dashboard' />
-        <CamPage path='/webcam' />
-      </Router>
-    </div>
-  );
+export default class App extends Component {
+  state = {}
+  render() {
+    return (
+      <div className="App">
+        <Header />
+        <Router>
+          <LoginPage path='/' />
+          <Dashboard path='/dashboard' />
+          <CamPage path='/webcam' startVideo={startVideo}/>
+        </Router>
+      </div>
+    )
+  }
 }
-
-export default App;
